Restructure cipher imports + separate AEAD for #29
