--- cipher-core/src/main/scala/tsec/cipher/common/mode/CBC.scala
@@ -1,8 +1,0 @@
-package tsec.cipher.common.mode
-
-sealed trait CBC
-
-/** our cbc mode takes 16 byte IVs
-  * https://crypto.stackexchange.com/questions/2594/initialization-vector-length-insufficient-in-aes
-  */
-object CBC extends DefaultModeKeySpec[CBC]("CBC", 16)
