--- cipher-core/src/main/scala/tsec/cipher/common/mode/CFB.scala
@@ -1,4 +1,0 @@
-package tsec.cipher.common.mode
-
-sealed trait CFB
-object CFB extends DefaultModeKeySpec[CFB]("CFB", 32)
