--- cipher-core/src/main/scala/tsec/cipher/common/mode/CFBx.scala
@@ -1,4 +1,0 @@
-package tsec.cipher.common.mode
-
-sealed trait CFBx
-object CFBx extends DefaultModeKeySpec[CFBx]("CFBx", 32)
