--- cipher-core/src/main/scala/tsec/cipher/common/mode/CTR.scala
@@ -1,4 +1,0 @@
-package tsec.cipher.common.mode
-
-sealed trait CTR
-object CTR extends DefaultModeKeySpec[CTR]("CTR", 16)
