--- cipher-core/src/main/scala/tsec/cipher/common/mode/CTS.scala
@@ -1,4 +1,0 @@
-package tsec.cipher.common.mode
-
-/** Will tentatively not implement */
-sealed trait CTS
