--- cipher-core/src/main/scala/tsec/cipher/common/mode/ECB.scala
@@ -1,4 +1,0 @@
-package tsec.cipher.common.mode
-
-trait ECB
-object ECB extends DefaultModeKeySpec[ECB]("ECB", 0)
