--- cipher-core/src/main/scala/tsec/cipher/common/mode/NoMode.scala
@@ -1,4 +1,0 @@
-package tsec.cipher.common.mode
-
-sealed trait NoMode
-object NoMode extends DefaultModeKeySpec[NoMode]("NONE", 0)
