--- cipher-core/src/main/scala/tsec/cipher/common/mode/OFB.scala
@@ -1,4 +1,0 @@
-package tsec.cipher.common.mode
-
-sealed trait OFB
-object OFB extends DefaultModeKeySpec[OFB]("OFB", 16)
