--- cipher-core/src/main/scala/tsec/cipher/common/mode/OFBx.scala
@@ -1,4 +1,0 @@
-package tsec.cipher.common.mode
-
-sealed trait OFBx
-object OFBx extends DefaultModeKeySpec[OFBx]("OFBx", 16)
