--- cipher-core/src/main/scala/tsec/cipher/common/mode/PCBC.scala
@@ -1,4 +1,0 @@
-package tsec.cipher.common.mode
-
-sealed trait PCBC
-object PCBC extends DefaultModeKeySpec[PCBC]("PCBC", 16)
