--- cipher-symmetric/src/main/scala/tsec/cipher/symmetric/core/package.scala
@@ -1,3 +1,0 @@
-package tsec.cipher.symmetric
-
-package object core {}
