+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/AES128.scala
@@ -1,4 +1,4 @@
 package tsec.cipher.symmetric.imports
 
 sealed trait AES128
-object AES128 extends WithSymmetricGenerator[AES128]("AES", 128)
+object AES128 extends WithAEADCipher[AES128]("AES", 128)
