+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/AES192.scala
@@ -1,4 +1,4 @@
 package tsec.cipher.symmetric.imports
 
 sealed trait AES192
-object AES192 extends WithSymmetricGenerator[AES192]("AES", 192)
+object AES192 extends WithAEADCipher[AES192]("AES", 192)
