+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/AES256.scala
@@ -1,4 +1,4 @@
 package tsec.cipher.symmetric.imports
 
 sealed trait AES256
-object AES256 extends WithSymmetricGenerator[AES256]("AES", 256)
+object AES256 extends WithAEADCipher[AES256]("AES", 256)
