+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/AESWrap.scala
@@ -1,4 +1,4 @@
 package tsec.cipher.symmetric.imports
 
 sealed trait AESWrap
-object AESWrap extends WithSymmetricGenerator[AESWrap]("AESWrap", 128)
+object AESWrap extends WithAEADCipher[AESWrap]("AESWrap", 128)
