+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/Blowfish.scala
@@ -1,4 +1,4 @@
 package tsec.cipher.symmetric.imports
 
 sealed trait Blowfish
-object Blowfish extends WithSymmetricGenerator[Blowfish]("Blowfish", 448)
+object Blowfish extends WithAEADCipher[Blowfish]("Blowfish", 448)
