+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/DES.scala
@@ -1,4 +1,4 @@
 package tsec.cipher.symmetric.imports
 
 sealed trait DES
-object DES extends WithSymmetricGenerator[DES]("DES", 56)
+object DES extends WithAEADCipher[DES]("DES", 56)
