+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/DESede.scala
@@ -1,4 +1,4 @@
 package tsec.cipher.symmetric.imports
 
 sealed trait DESede
-object DESede extends WithSymmetricGenerator[DESede]("DESede", 168)
+object DESede extends WithAEADCipher[DESede]("DESede", 168)
