+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/DESedeWrap.scala
@@ -1,4 +1,4 @@
 package tsec.cipher.symmetric.imports
 
 sealed trait DESedeWrap
-object DESedeWrap extends WithSymmetricGenerator[DESedeWrap]("DESedeWrap", 164)
+object DESedeWrap extends WithAEADCipher[DESedeWrap]("DESedeWrap", 164)
