+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/DefaultEncryptor.scala
@@ -1,11 +1,10 @@
 package tsec.cipher.symmetric.imports
 
-import tsec.cipher.common._
-import tsec.cipher.common.mode.CTR
+import tsec.cipher.symmetric._
 import tsec.cipher.common.padding.NoPadding
 
 
-sealed abstract class Encryptor[A: SymmetricAlgorithm] {
+sealed abstract class Encryptor[A: SymmetricCipher] {
   lazy val instance: Either[NoSuchInstanceError, JCASymmetricCipher[A, CTR, NoPadding]] =
     JCASymmetricCipher[A, CTR, NoPadding]
