+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/aead/JCAAEAD.scala
@@ -1,52 +1,30 @@
-package tsec.cipher.symmetric.imports.threadlocal
+package tsec.cipher.symmetric.imports.aead
 
-import java.util.{ArrayDeque => JQueue}
 import javax.crypto.{Cipher => JCipher}
 import cats.syntax.either._
-import tsec.cipher.common._
 import tsec.cipher.symmetric._
+import tsec.cipher.symmetric.mode._
+import tsec.cipher.symmetric.imports._
 import tsec.cipher.common.padding.Padding
 import tsec.cipher.symmetric.SymmetricCipherAlgebra
-import tsec.cipher.symmetric.imports.{SecretKey, SymmetricCipher}
-import tsec.cipher.symmetric.mode._
 import tsec.common.ErrorConstruct._
 
-abstract class JCATLSymmetric[A, M, P](
-    implicit algoTag: SymmetricCipher[A],
-    modeSpec: CipherMode[M],
+class JCAAEAD[A, M, P](
+    implicit algoTag: AEADCipher[A],
+    modeSpec: AEADMode[M],
     paddingTag: Padding[P]
 ) extends SymmetricCipherAlgebra[Either[CipherError, ?], A, M, P, SecretKey] {
 
   type C = JCipher
 
-  protected val local: ThreadLocal[JQueue[JCipher]]
-
-  private def catchGen: Either[InstanceInitError, JCipher] =
+  def genInstance: Either[CipherError, JCipher] =
     Either
       .catchNonFatal(JCipher.getInstance(s"${algoTag.algorithm}/${modeSpec.algorithm}/${paddingTag.algorithm}"))
       .mapError(InstanceInitError.apply)
 
-  def genInstance: Either[CipherError, JCipher] =
-    Either
-      .catchNonFatal {
-        val threadLocal = local.get()
-        threadLocal.poll()
-      }
-      .flatMap { inst =>
-        if (inst == null)
-          catchGen
-        else
-          Right(inst)
-      }
-      .mapError(InstanceInitError.apply)
-
-  def replace(instance: JCipher): Either[DecryptError, Unit] =
-    Right(local.get().add(instance))
-
-  /** Stateful operations for internal use
-    * Made private so as to not encourage any use of stateful operations
+  /** Stateful operations for internal use Made private so as to not encourage any use of stateful operations.
     * The only other option would be to defer these operations with something like IO, given they are stateful
-   */
+    */
   protected[symmetric] def initEncryptor(
       e: JCipher,
       secretKey: SecretKey[A]
@@ -64,14 +42,18 @@ abstract class JCATLSymmetric[A, M, P](
   ): Either[CipherKeyError, Unit] =
     Either
       .catchNonFatal({
-        decryptor.init(JCipher.DECRYPT_MODE, SecretKey.toJavaKey[A](key), ParameterSpec.toRepr[M](modeSpec.buildIvFromBytes(iv)))
+        decryptor.init(
+          JCipher.DECRYPT_MODE,
+          SecretKey.toJavaKey[A](key),
+          ParameterSpec.toRepr[M](modeSpec.buildIvFromBytes(iv))
+        )
       })
       .mapError(CipherKeyError.apply)
 
   protected[symmetric] def setAAD(e: JCipher, aad: AAD): Either[CipherKeyError, Unit] =
     Either.catchNonFatal(e.updateAAD(aad.aad)).mapError(CipherKeyError.apply)
-  /** End stateful ops */
 
+  /** End stateful ops  */
   /** Encrypt our plaintext with a tagged secret key
     *
     * @param plainText the plaintext to encrypt
@@ -89,7 +71,6 @@ abstract class JCATLSymmetric[A, M, P](
         .catchNonFatal(instance.doFinal(plainText.content))
         .mapError(EncryptError.apply)
       iv <- Either.fromOption(Option(instance.getIV), IvError("No IV found"))
-      _  <- replace(instance)
     } yield CipherText(encrypted, iv)
 
   /** Encrypt our plaintext using additional authentication parameters,
@@ -114,7 +95,6 @@ abstract class JCATLSymmetric[A, M, P](
         .catchNonFatal(instance.doFinal(plainText.content))
         .mapError(EncryptError.apply)
       iv <- Either.fromOption(Option(instance.getIV), IvError("No IV found"))
-      _  <- replace(instance)
     } yield CipherText(encrypted, iv)
 
   /** Decrypt our ciphertext
@@ -133,7 +113,6 @@ abstract class JCATLSymmetric[A, M, P](
       decrypted <- Either
         .catchNonFatal(instance.doFinal(cipherText.content))
         .mapError(DecryptError.apply)
-      _ <- replace(instance)
     } yield PlainText(decrypted)
 
   /** Decrypt our ciphertext using additional authentication parameters,
@@ -157,65 +136,26 @@ abstract class JCATLSymmetric[A, M, P](
       decrypted <- Either
         .catchNonFatal(instance.doFinal(cipherText.content))
         .mapError(DecryptError.apply)
-      _ <- replace(instance)
     } yield PlainText(decrypted)
 }
 
-object JCATLSymmetric {
-
-  protected[imports] def getJCipherUnsafe[A, M, P](
-      implicit algoTag: SymmetricCipher[A],
-      modeSpec: CipherMode[M],
-      paddingTag: Padding[P]
-  ): JCipher = JCipher.getInstance(s"${algoTag.algorithm}/${modeSpec.algorithm}/${paddingTag.algorithm}")
-
-  /** generate queue unsafe
-    *
-    * @param queueLen
-    * @tparam A
-    * @tparam M
-    * @tparam P
-    * @return
-    */
-  protected[imports] def genQueueUnsafe[A: SymmetricCipher, M: CipherMode, P: Padding](
-      queueLen: Int
-  ): JQueue[JCipher] = {
-    val q = new JQueue[JCipher]()
-    (0 until queueLen)
-      .foreach(
-        _ => q.add(getJCipherUnsafe)
-      )
-    q
-  }
+object JCAAEAD {
 
   /** Attempt to initialize an instance of the cipher with the given type parameters
-    * All processing is done on threadlocal, to guarantee no leaked instances
-    * @param queueLen the length of the queue
+    * If the cipher doesn't exist/is not supported, it will return NoSuchIntanceError
+    *
     * @tparam A Symmetric Cipher Algorithm
     * @tparam M Mode of operation
     * @tparam P Padding mode
     * @return
     */
-  def apply[A: SymmetricCipher, M: CipherMode, P: Padding](
-      queueLen: Int = 15
-  ): Either[NoSuchInstanceError.type, JCATLSymmetric[A, M, P]] =
-    for {
-      q <- Either.catchNonFatal(genQueueUnsafe(queueLen)).leftMap(_ => NoSuchInstanceError)
-      tl <- Either
-        .catchNonFatal {
-          new ThreadLocal[JQueue[JCipher]] {
-            override def initialValue(): JQueue[JCipher] =
-              q
-          }
-        }
-        .leftMap(_ => NoSuchInstanceError)
-    } yield
-      new JCATLSymmetric[A, M, P] {
-        protected val local: ThreadLocal[JQueue[JCipher]] = tl
-      }
+  def apply[A: AEADCipher, M: AEADMode, P: Padding]: Either[NoSuchInstanceError.type, JCAAEAD[A, M, P]] = {
+    val c = new JCAAEAD[A, M, P]
+    c.genInstance.map(_ => c).leftMap(_ => NoSuchInstanceError)
+  }
 
-  def genInstance[A: SymmetricCipher, M: CipherMode, P: Padding]
-    : Either[NoSuchInstanceError.type, JCATLSymmetric[A, M, P]] = apply[A, M, P]()
+  implicit def genSym[A: AEADCipher, M: AEADMode, P: Padding]: Either[NoSuchInstanceError.type, JCAAEAD[A, M, P]] =
+    apply[A, M, P]
 
   /** ┌(▀Ĺ̯▀)–︻╦╤─ "You will never get away with an unsafe instance!!"
     *
@@ -226,13 +166,7 @@ object JCATLSymmetric {
     * @tparam P Padding mode
     * @return
     */
-  def getCipherUnsafe[A: SymmetricCipher, M: CipherMode, P: Padding](queueLen: Int): JCATLSymmetric[A, M, P] = {
-    val queue = genQueueUnsafe(queueLen)
-    new JCATLSymmetric[A, M, P] {
-      protected val local: ThreadLocal[JQueue[JCipher]] = new ThreadLocal[JQueue[JCipher]] {
-        override def initialValue(): JQueue[JCipher] = queue
-      }
-    }
-  }
+  def getCipherUnsafe[A: AEADCipher, M: AEADMode, P: Padding]: JCAAEAD[A, M, P] =
+    new JCAAEAD[A, M, P]
 
 }
