+++ cipher-symmetric/src/main/scala/tsec/cipher/symmetric/imports/aead/JCAAEADPure.scala
@@ -1,20 +1,17 @@
-package tsec.cipher.symmetric.imports.experimental
+package tsec.cipher.symmetric.imports.aead
 
-import java.util.concurrent.{ConcurrentLinkedQueue => JQueue}
+import cats.effect.Sync
 import javax.crypto.{Cipher => JCipher}
-
-import cats.effect.{IO, Sync}
-import cats.syntax.flatMap._
-import cats.syntax.functor._
-import tsec.cipher.common._
-import tsec.cipher.common.mode._
+import tsec.cipher.symmetric.mode._
 import tsec.cipher.common.padding.Padding
-import tsec.cipher.symmetric.core.SymmetricCipherAlgebra
-import tsec.cipher.symmetric.imports.{SecretKey, SymmetricAlgorithm}
+import java.util.concurrent.{ConcurrentLinkedQueue => JQueue}
+import tsec.cipher.symmetric._
+import tsec.cipher.symmetric.imports._
+import cats.syntax.all._
 
-sealed abstract class JCASymmPure[F[_], A, M, P](queue: JQueue[JCipher])(
-    implicit algoTag: SymmetricAlgorithm[A],
-    modeSpec: ModeKeySpec[M],
+class JCAAEADPure[F[_], A, M, P](queue: JQueue[JCipher])(
+    implicit algoTag: AEADCipher[A],
+    modeSpec: AEADMode[M],
     paddingTag: Padding[P],
     F: Sync[F]
 ) extends SymmetricCipherAlgebra[F, A, M, P, SecretKey] {
@@ -26,7 +23,7 @@ sealed abstract class JCASymmPure[F[_], A, M, P](queue: JQueue[JCipher])(
     if (inst != null)
       inst
     else
-      JCASymmPure.getJCipherUnsafe[A, M, P]
+      JCAAEADPure.getJCipherUnsafe[A, M, P]
   }
 
   def replace(instance: JCipher): F[Boolean] =
@@ -37,19 +34,24 @@ sealed abstract class JCASymmPure[F[_], A, M, P](queue: JQueue[JCipher])(
       instance: JCipher,
       secretKey: SecretKey[A]
   ): F[Unit] =
-    F.delay(instance.init(JCipher.ENCRYPT_MODE, SecretKey.toJavaKey[A](secretKey), ParameterSpec.toRepr[M](modeSpec.genIv)))
+    F.delay(
+      instance.init(JCipher.ENCRYPT_MODE, SecretKey.toJavaKey[A](secretKey), ParameterSpec.toRepr[M](modeSpec.genIv))
+    )
 
   protected[symmetric] def initDecryptor(
       instance: JCipher,
       key: SecretKey[A],
       iv: Array[Byte]
   ): F[Unit] =
-    F.delay(instance.init(JCipher.DECRYPT_MODE, SecretKey.toJavaKey[A](key), ParameterSpec.toRepr[M](modeSpec.buildIvFromBytes(iv))))
+    F.delay(
+      instance
+        .init(JCipher.DECRYPT_MODE, SecretKey.toJavaKey[A](key), ParameterSpec.toRepr[M](modeSpec.buildIvFromBytes(iv)))
+    )
 
   protected[symmetric] def setAAD(e: JCipher, aad: AAD): F[Unit] =
     F.delay(e.updateAAD(aad.aad))
-  /** End stateful ops */
 
+  /** End stateful ops */
   /** Encrypt our plaintext with a tagged secret key
     *
     * @param plainText the plaintext to encrypt
@@ -131,11 +133,11 @@ sealed abstract class JCASymmPure[F[_], A, M, P](queue: JQueue[JCipher])(
     } yield PlainText(decrypted)
 }
 
-object JCASymmPure {
+object JCAAEADPure {
 
   protected[imports] def getJCipherUnsafe[A, M, P](
-      implicit algoTag: SymmetricAlgorithm[A],
-      modeSpec: ModeKeySpec[M],
+      implicit algoTag: AEADCipher[A],
+      modeSpec: AEADMode[M],
       paddingTag: Padding[P]
   ): JCipher = JCipher.getInstance(s"${algoTag.algorithm}/${modeSpec.algorithm}/${paddingTag.algorithm}")
 
@@ -147,7 +149,7 @@ object JCASymmPure {
     * @tparam P
     * @return
     */
-  protected[imports] def genQueueUnsafe[A: SymmetricAlgorithm, M: ModeKeySpec, P: Padding](
+  protected[imports] def genQueueUnsafe[A: AEADCipher, M: AEADMode, P: Padding](
       queueLen: Int
   ): JQueue[JCipher] = {
     val q = new JQueue[JCipher]()
@@ -166,14 +168,14 @@ object JCASymmPure {
     * @tparam P Padding mode
     * @return
     */
-  def apply[F[_], A: SymmetricAlgorithm, M: ModeKeySpec, P: Padding](
+  def apply[F[_], A: AEADCipher, M: AEADMode, P: Padding](
       queueLen: Int = 15
-  )(implicit F: Sync[F]): F[JCASymmPure[F, A, M, P]] =
+  )(implicit F: Sync[F]): F[JCAAEADPure[F, A, M, P]] =
     for {
       q <- F.delay(genQueueUnsafe(queueLen))
-    } yield new JCASymmPure[F, A, M, P](q) {}
+    } yield new JCAAEADPure[F, A, M, P](q) {}
 
-  implicit def genInstance[F[_]: Sync, A: SymmetricAlgorithm, M: ModeKeySpec, P: Padding]: F[JCASymmPure[F, A, M, P]] =
+  implicit def genInstance[F[_]: Sync, A: AEADCipher, M: AEADMode, P: Padding]: F[JCAAEADPure[F, A, M, P]] =
     apply[F, A, M, P]()
 
 }
